==> Code/01_StableClimExtract.Rmd <==
---
title: "01 - StableClim Extract"
author: "Stuart C Brown"
output: 
  html_document: 
    theme: flatly
    toc: yes
    toc_depth: 3
    toc_float: true
    fig_width: 7
    fig_height: 5
    fig_caption: yes
    number_sections: no
    keep_md: yes
    df_print: paged
    code_folding: show
    highlight: tango
bibliography: references.bib
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, rows.print = 10, cols.print = 10,
                      rownames.print = FALSE, warning = FALSE,
                      message = FALSE, eval = TRUE, results = "markup", 
                      error = FALSE, strip.white = TRUE,
                      tidy = "styler", dpi = 320, fig.align = "center",
                      strip.white = TRUE, 
                      dev.args = list(png  = list(type = "cairo")))
if(!dir.exists("../outputs")) {
  dir.create("../outputs/", recursive = TRUE)
}
```

## Extract data from StableClim

We need to extract the climate data from the StableClim (@brown2020) database. The data for StableClim can be downloaded from [here](https://doi.org/10.25909/5ea59831121bc).

## Load libraries

In this section we load the necessary libraries, set the paths to the StableClim database, and import a land/sea mask for plotting.

```{r setSC, eval = TRUE, echo = TRUE}
# load libs
library(terra) ## 1.5.21
library(data.table)
library(pbapply)
library(sf) ## 1.0-5
library(rnaturalearth)
library(rnaturalearthdata)
library(ggplot2)
library(scales)
library(cptcity)
library(MetBrewer)

# StableClim path
scPath <- "/mnt/c/StableClim"

# download natural earth data function
source("source/ne_download2.R")

# Download ocean shapefile to use as a mask
## if already downloaded and processed, read in ocean mask raster
if (!file.exists("outputs/ocean_mask.grd")) {
  ocean_mask <- ne_download2(scale = 50, type = "ocean", category = "physical",
                          destdir = file.path("outputs"), load = TRUE,
                          returnclass = "sf")
  st_write(ocean_mask, "outputs/output_shapes.gpkg", layer = "ne_ocean", quiet = TRUE)
  shp <- list.files("outputs/", "ne_50m_ocean")
  invisible(file.remove(file.path("outputs", shp)))
  ocean_mask <- rasterize(x = vect(ocean_mask), 
                        y = rast(resolution = 2.5, crs = "EPSG:4326",
                                 extent = c(-180,180,-90,90)), 
                        cover = TRUE)
  ocean_mask[ocean_mask < 0.85] <- NA
  ocean_mask[!is.na(ocean_mask)] <- 1
  # need to open up the mediterranean sea to the Atlantic
  cell_idx <- cellFromXY(ocean_mask, 
                         rbind(cbind(seq(-6, 1.5, by = 2.5), 36),
                               cbind(8.5, 38.75),
                               cbind(11.5, 36),
                               cbind(11.5, 34)))
  ocean_mask[cell_idx] <- 1
  writeRaster(ocean_mask, filename = "outputs/ocean_mask.grd",
            datatype = "INT1U", overwrite = TRUE)
  } else {
    ocean_mask <- rast("outputs/ocean_mask.grd")
  }
```

## Load StableClim data

Here we will load in the threshold and regression data from StableClim

```{r loadSC}
# Pre-industrial thresholds
pithresh <- readRDS(file.path(scPath, "StableClim_piControl_thresholds.RDS"))

# Paleo regressions
pal_regr <- readRDS(file.path(scPath, "StableClim_past_RegionalTemperatureRegressions.RDS"))
pal_regr_ocean <- pal_regr$Global.Ocean

# Historical and future regressions
rcp_regr <- readRDS(file.path(scPath, "StableClim_HistoricalRCP_RegionalTemperatureRegressions.RDS"))
rcp45_ocean <- rcp_regr$rcp45.Global.Ocean
rcp85_ocean <- rcp_regr$rcp85.Global.Ocean

# Import the thresholds for each scenario
## Warming Ocean
ocn_warm <- setDT(pithresh$`Warm periods`)
ocn_warm <- ocn_warm[RegionType == "Global" & Region == "Ocean",][["95%"]]

## Ocean Absolute
ocn_abso <- setDT(pithresh$`All Periods [absolute]`)
ocn_abso <- ocn_abso[RegionType == "Global" & Region == "Ocean",][["95%"]]

thresh <- list(ocn_warm, ocn_abso)
names(thresh) <- c("ocn_warm", "ocn_abso")
```

From the StableClim data, the threshold for defining extreme natural warming events is `r round(ocn_warm, 5) * 100` °C/century if we consider only warming periods, and slightly less at `r round(ocn_abso, 5) * 100` °C/century if we take the absolute values of all time periods.

## Subset the RCP data to extreme periods

Now we can subset the Paleo period and RCP 4.5 and 8.5 to periods where global warming is greater than the natural warming thresholds extracted above.

```{r subsetRCP}
# Subset the paleo regressions using the different thresholds
pal_subset <- lapply(names(thresh), function(x) {
  t <- thresh[x]
  sub <- pal_regr_ocean[Slope >= t, ]
  return(sub)
  })
names(pal_subset) <- names(thresh)

# Subset the rcp regressions using the different threshold
rcp45_subset <- lapply(names(thresh), function(x) {
  t <- thresh[x]
  sub <- rcp45_ocean[Slope >= t, ]
  return(sub)
}
)
names(rcp45_subset) <- names(thresh)

rcp85_subset <- lapply(names(thresh), function(x) {
  t <- thresh[x]
  sub <- rcp85_ocean[Slope >= t, ]
  return(sub)
}
)
names(rcp85_subset) <- names(thresh)
```

Due to difference in the number of models in the ensembles for RCP 4.5 and RCP 8.5, the historical periods that are selected differ slightly.

## Load raster data

Now we have to load the raster data contained in StableClim which holds the trend and variability for temperatures.

```{r loadSCRas}
# Load in the raster stacks and subset to extreme periods
# Paleo
pal_rast_trend <- rast(file.path(scPath, "ncdf/regressions/StableClim_Regression_past_ts.nc"), 
                       subds = "ts_trend")
pal_rast_var <- rast(file.path(scPath, "ncdf/regressions/StableClim_Regression_past_ts.nc"), 
                     subds = "ts_variability")

## indices of raster layers where the piControl threshold is exceeded
zInd <- as.integer(sapply(strsplit(names(pal_rast_trend), "="), "[", 2))
pal_zExceed <- lapply(pal_subset, function(x) {
  yr <- x$Start
  idx <- which(zInd %in% yr)
})

# RCP 4.5
rcp45_rast_trend <- rast(file.path(scPath, "ncdf/regressions/StableClim_Regression_rcp45_ts.nc"), 
                         subds = "ts_trend")
rcp45_rast_var <- rast(file.path(scPath, "ncdf/regressions/StableClim_Regression_rcp45_ts.nc"), 
                       subds = "ts_variability")
zInd <- as.integer(sapply(strsplit(names(rcp45_rast_trend), "="), "[", 2))
rcp45_zExceed <- lapply(rcp45_subset, function(x) {
  yr <- x$Start
  idx <- which(zInd %in% yr)
})

# RCP 8.5
rcp85_rast_trend <- rast(file.path(scPath, "ncdf/regressions/StableClim_Regression_rcp85_ts.nc"), 
                         subds = "ts_trend")
rcp85_rast_var <- rast(file.path(scPath, "ncdf/regressions/StableClim_Regression_rcp85_ts.nc"), 
                       subds = "ts_variability")
rcp85_zExceed <- lapply(rcp85_subset, function(x) {
  yr <- x$Start
  idx <- which(zInd %in% yr)
})

```

## Spatiotemporal means from extreme periods

Generate spatiotemporal means during warming periods in the ocean

```{r stMeanOcean}
# Paleo
pal_rast_sub <- lapply(names(pal_zExceed), function(x) {
  sub <- pal_zExceed[[x]]
  trend <- pal_rast_trend[[sub]]
  var <- pal_rast_var[[sub]]
  # spatiotemporal means
  meanTrend <- app(trend, mean)
  meanVar <- app(var, mean)
  meanSNR <- abs(meanTrend)/meanVar
  # global extreme periods
  stTrend <- pal_subset[[x]]
  stTrend <- stTrend[, .(Slope = mean(Slope))][["Slope"]]
  # standardised trend and SNR
  trend <- abs(meanTrend)/stTrend
  stSNR <- trend/meanVar
  s <- rast(list(meanTrend, trend, meanVar, meanSNR, stSNR))
  names(s) <- paste0(c("Trend_", "stTrend_", "Var_", "SNR_", "stSNR_"), x)
  return(s)
})
names(pal_rast_sub) <- names(pal_zExceed)

# Slope for standardising the RCP data
stSlope <- mean(c(rcp45_subset$ocn_warm$Slope,
                  rcp85_subset$ocn_warm$Slope))

# RCP 4.5
rcp45_rast_sub <- lapply(names(rcp45_zExceed), function(x) {
  sub <- rcp45_zExceed[[x]]
  trend <- rcp45_rast_trend[[sub]]
  var <- rcp45_rast_var[[sub]]
  meanTrend <- app(trend, mean)
  meanVar <- app(var, mean)
  meanSNR <- abs(meanTrend)/meanVar
  trend <- abs(meanTrend)/stSlope
  stSNR <- trend/meanVar
  s <- rast(list(meanTrend, trend, meanVar, meanSNR, stSNR))
  names(s) <- paste0(c("Trend_", "stTrend_", "Var_", "SNR_", "stSNR_"), x)
  return(s)
})
names(rcp45_rast_sub) <- names(rcp45_zExceed)

# RCP 8.5
rcp85_rast_sub <- lapply(names(rcp85_zExceed), function(x) {
  sub <- rcp85_zExceed[[x]]
  trend <- rcp85_rast_trend[[sub]]
  var <- rcp85_rast_var[[sub]]
  meanTrend <- app(trend, mean)
  meanVar <- app(var, mean)
  meanSNR <- abs(meanTrend)/meanVar
  trend <- abs(meanTrend)/stSlope
  stSNR <- trend/meanVar
  s <- rast(list(meanTrend, trend, meanVar, meanSNR, stSNR))
  names(s) <- paste0(c("Trend_", "stTrend_", "Var_", "SNR_", "stSNR_"), x)
  return(s)
})
names(rcp85_rast_sub) <- names(rcp85_zExceed)

snr_ras <- rast(list(pal_rast_sub$ocn_warm$stSNR_ocn_warm, 
                     rcp45_rast_sub$ocn_warm$stSNR_ocn_warm,
                     rcp85_rast_sub$ocn_warm$stSNR_ocn_warm))
```

Looking at the SNR rasters, we see that under RCP 8.5, SNR values are much higher through the tropics and mid-latitudes, but lower in the Atlantic - particularly south of Greenland. In this plot, warmer colours show higher SNR values.

```{r plotSNR, echo = FALSE, eval = TRUE, fig.show='asis', fig.cap="SNR values for the past and two future scenarios", fig.align='center', fig.height=5, fig.width=9}
land <- ne_coastline(50, returnclass = "sf")
x <- mask(snr_ras, ocean_mask)
names(x) <- c("Paleo", "RCP 4.5", "RCP 8.5")
xdf <- melt(setDT(as.data.frame(x, xy = TRUE)), id.vars = c("x", "y"))

ggplot() +
  geom_raster(data = xdf, aes(x, y, fill = value)) +
  facet_wrap(~variable, ncol = 2) +
  scale_fill_gradientn(colours = met.brewer(name = "Hokusai1", n = 100, direction = -1),
                       limits = c(0, 6),
                       labels = c("Low","", "High"),
                       breaks = c(0, 3, 6),
                       oob = squish,
                       guide = guide_colourbar(
                         title = "SNR",
                         title.position = "top",
                         title.hjust = 0.5,
                         title.theme = element_text(size = 12, colour = "black"),
                         label.position = "bottom",
                         label.hjust = 0.5,
                         label.theme = element_text(size = 12, colour = "black"),
                         barwidth = 10,
                         barheight = 0.5,
                         nbin = 100,
                         frame.colour = "black",
                         frame.linewidth = 0.5
                       )) +
  geom_sf(data = land, inherit.aes = FALSE) +
  coord_sf(xlim = c(-180,180), ylim = c(-90,90),
           expand = FALSE) +
  cowplot::theme_map() +
  theme(legend.position = c(0.75, 0.25),
        legend.justification = "center",
        legend.direction = "horizontal",
        strip.text.x = element_text(size = 12, colour = "black"))
```

## Export the rasters

Now we write the rasters out. Output is `{terra}/{raster}` native raster format.

```{r saveSCRast}
# Save rasters
writeRaster(pal_rast_sub$ocn_warm,
            filename = "outputs/paleo_rasters_ocean_warming.grd",
            datatype = "FLT4S")
writeRaster(rcp45_rast_sub$ocn_warm,
            filename = "outputs/rcp45_rasters_ocean_warming.grd",
            datatype = "FLT4S")
writeRaster(rcp85_rast_sub$ocn_warm,
            filename = "outputs/rcp85_rasters_ocean_warming.grd",
            datatype = "FLT4S")
```

## Delta SNR

Now we can use the extracted data to calculate Delta SNR between the past and the future.

```{r calcDelSNR}
snr_deltaRCP45 <- lapply(c("ocn_warm", "ocn_abso"), function(x) {
  ## 5th layer is standardised SNR
  s <- rast(list(pal_rast_sub[[x]][[5]], rcp45_rast_sub[[x]][[5]]))
  sOcean <- mask(s, ocean_mask)
  minS <- min(global(sOcean, min, na.rm = TRUE))
  maxS <- max(global(sOcean, max, na.rm = TRUE))
  sOcean <- stretch(sOcean, minv = 0, maxv = 1, smin = minS, smax = maxS)
  names(sOcean) <- c("past", "future")
  sLand <- mask(s, ocean_mask, inverse = TRUE)
  minS <- min(global(sLand, min, na.rm = TRUE))
  maxS <- max(global(sLand, max, na.rm = TRUE))
  sLand <- stretch(sLand, minv = 0, maxv = 1, smin = minS, smax = maxS)
  names(sLand) <- c("past", "future")
  rExt <- rast(resolution = 2.5, nlyrs = 2, crs = "EPSG:4326")
  values(rExt) <- 0
  values(rExt[[1]]) <- values(cover(sOcean[["past"]], sLand[["past"]]))
  values(rExt[[2]]) <- values(cover(sOcean[["future"]], sLand[["future"]]))
  names(rExt) <- c("past", "future")
  rDel <- rExt[["future"]] - rExt[["past"]]
  names(rDel) <- paste0("Delta_stSNR_", x)
  return(rDel)
})
snr_deltaRCP45 <- rast(snr_deltaRCP45)

snr_deltaRCP85 <- lapply(c("ocn_warm", "ocn_abso"), function(x) {
  s <- rast(list(pal_rast_sub[[x]][[5]], rcp85_rast_sub[[x]][[5]]))
  sOcean <- mask(s, ocean_mask)
  minS <- min(global(sOcean, min, na.rm = TRUE))
  maxS <- max(global(sOcean, max, na.rm = TRUE))
  sOcean <- stretch(sOcean, minv = 0, maxv = 1, smin = minS, smax = maxS)
  names(sOcean) <- c("past", "future")
  sLand <- mask(s, ocean_mask, inverse = TRUE)
  minS <- min(global(sLand, min, na.rm = TRUE))
  maxS <- max(global(sLand, max, na.rm = TRUE))
  sLand <- stretch(sLand, minv = 0, maxv = 1, smin = minS, smax = maxS)
  names(sLand) <- c("past", "future")
  rExt <- rast(resolution = 2.5, nlyrs = 2, crs = "EPSG:4326")
  values(rExt) <- 0
  values(rExt[[1]]) <- values(cover(sOcean[["past"]], sLand[["past"]]))
  values(rExt[[2]]) <- values(cover(sOcean[["future"]], sLand[["future"]]))
  names(rExt) <- c("past", "future")
  rDel <- rExt[["future"]] - rExt[["past"]]
  names(rDel) <- paste0("Delta_stSNR_", x)
  return(rDel)
})
snr_deltaRCP85 <- rast(snr_deltaRCP85)

writeRaster(snr_deltaRCP45[[1]], "outputs/snr_st_delta_rcp45.grd")
writeRaster(snr_deltaRCP85[[1]], "outputs/snr_st_delta_rcp85.grd")

```

Plots of Delta SNR, show that SNR is higher in the future for most of the world with the exception of the Atlantic.

```{r plotDelSNR, echo = FALSE, fig.cap="Negative values = higher SNR in the past", fig.align='center', fig.height=3.5, fig.width=9}
x <- mask(rast(list(snr_deltaRCP45[["Delta_stSNR_ocn_warm"]],
                    snr_deltaRCP85[["Delta_stSNR_ocn_warm"]])), 
          mask = ocean_mask)
names(x) <- c("RCP 4.5", "RCP 8.5")
xdf <- melt(setDT(as.data.frame(x, xy = TRUE)), id.vars = c("x", "y"))
ggplot() +
  geom_raster(data = xdf, aes(x, y, fill = value)) +
  facet_wrap(~variable, ncol = 2) +
  scale_fill_gradientn(colours = cpt("ncl_BlWhRe", n = 100),
                       limits = c(-1, 1),
                       labels = c("Lower","No change", "Higher"),
                       breaks = c(-1, 0, 1),
                       guide = guide_colourbar(
                         title = "Delta SNR",
                         title.position = "top",
                         title.hjust = 0.5,
                         title.theme = element_text(size = 12, colour = "black"),
                         label.position = "bottom",
                         label.hjust = 0.5,
                         label.theme = element_text(size = 12, colour = "black"),
                         barwidth = 10,
                         barheight = 0.5,
                         nbin = 100,
                         frame.colour = "black",
                         frame.linewidth = 0.5
                       )) +
  geom_sf(data = land, inherit.aes = FALSE) +
  coord_sf(xlim = c(-180,180), ylim = c(-90,90),
           expand = FALSE) +
  cowplot::theme_map() +
  theme(legend.position = "bottom",
        legend.direction = "horizontal",
        legend.justification = "center",
        strip.text.x = element_text(size = 12, colour = "black"))
```

Delta SNR between the two SNR scenarios, shows that SNR is higher under RCP 8.5

```{r plotDelSNRRCP, echo = FALSE, fig.cap="SNR RCP 8.5 - SNR RCP 4.5",fig.align='center', fig.height=5, fig.width=8}
x <- x[["RCP 8.5"]] - x[["RCP 4.5"]]
names(x) <- "value"
xdf <- setDT(as.data.frame(x, xy = TRUE))
ggplot() +
  geom_raster(data = xdf, aes(x, y, fill = value)) +
  scale_fill_gradientn(colours = cpt("ocal_german_flag_smooth", n = 100, rev = TRUE),
                       limits = range(scales::pretty_breaks(5)(range(xdf$value))),
                       labels = scales::pretty_breaks(5)(range(xdf$value)),
                       breaks = scales::pretty_breaks(5)(range(xdf$value)),
                       guide = guide_colourbar(
                         title = "RCP 8.5 - RCP 4.5",
                         title.position = "top",
                         title.hjust = 0.5,
                         title.theme = element_text(size = 12, colour = "black"),
                         label.position = "bottom",
                         label.hjust = 0.5,
                         label.theme = element_text(size = 12, colour = "black"),
                         barwidth = 10,
                         barheight = 0.5,
                         nbin = 100,
                         frame.colour = "black",
                         frame.linewidth = 0.5
                       )) +
  geom_sf(data = land, inherit.aes = FALSE) +
  coord_sf(xlim = c(-180,180), ylim = c(-90,90),
           expand = FALSE) +
  cowplot::theme_map() +
  theme(legend.position = "bottom",
        legend.direction = "horizontal",
        legend.justification = "center",
        strip.text.x = element_text(size = 12, colour = "black"))
```

## References
