==> Code/02_SppRichness.Rmd <==
---
title: "02 - Marine Species Richness"
author: "Stuart Brown"
output: 
  html_document: 
    theme: flatly
    toc: yes
    toc_depth: 3
    toc_float: true
    fig_width: 7
    fig_height: 5
    fig_caption: yes
    number_sections: no
    keep_md: yes
    df_print: paged
    code_folding: show
    highlight: tango
bibliography: references.bib
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, rows.print = 10, cols.print = 10,
                      rownames.print = FALSE, warning = FALSE,
                      message = FALSE, eval = TRUE, results = "markup", 
                      error = FALSE, strip.white = TRUE,
                      tidy = "styler", dpi = 320, fig.align = "center",
                      strip.white = TRUE, 
                      dev.args = list(png  = list(type = "cairo")))
```

## Load marine species richness data

Here we load in the marine species richness data generated from AquaMaps, and the marine biogeographic realms.

We used marine species richness data extracted from [AquaMaps](www.aquamaps.org). AquaMaps provides relative probabilities of current-day species occurrence at 0.5° x 0.5° resolution based on ecological niche models and expert knowledge. We discarded distribution data for species with occurences in \< 10 grid cells (i.e., the criteria used by AquaMaps to define data-scarce species). This resulted in a total of 14,173 marine algae, plant, invertebrate, fish, mammal, and reptile species. A species richness raster was then generated by adding the total number of species in each grid cell (@garcíamolinos2015).

Marine biogeographic realms are taken from @costello2017. As of writing this document a shapefile of the biogeographic realms is availble [here](https://auckland.figshare.com/articles/GIS_shape_files_of_realm_maps/5596840).

```{r loadLibsAndData, results='markup'}
library(data.table)
library(terra)
library(sf)
library(ggplot2)
library(rnaturalearth)
library(DescTools)
library(scales)
library(paletteer)
library(viridis)

# load in the Istat function to compare species richness maps
source("source/Istat.R")

# Don't use spherical geometry
sf_use_s2(FALSE)

coastlines <- ne_coastline(50, returnclass = "sf")
oceans <- read_sf("outputs/output_shapes.gpkg", layer = "ne_ocean")

# Marine realms from Costello et al
## need to source and then change path!
## available from https://auckland.figshare.com/articles/GIS_shape_files_of_realm_maps/5596840
regions <- read_sf("/mnt/c/Shapes/MarineRealmsShapeFile/MarineRealms.shp")

st_agr(oceans) <- "constant"
st_agr(regions) <- "constant"

# Clip regions to oceans
regions <- st_transform(
  # project to plate-carree before intersecting
  # add 0 buffer to avoid TopologyException error
  st_intersection(st_buffer(st_transform(regions, crs = "EPSG:32662"), 0),
                  st_buffer(st_transform(oceans[2], crs = "EPSG:32662"), 0)),
  crs = "EPSG:4326")

st_write(regions, "outputs/output_shapes.gpkg", layer = "marine_regions_clipped",
         quiet = TRUE)

# 2.5° Ocean mask
ocean_mask <- rast("outputs/ocean_mask.grd")
```

## Sensitivity analysis for species richness data

We have produced maps of marine species richness at thresholds between 0 and 0.5 at 0.1 increments. Here we perform some simple tests to see if changing the thresholds causes significant differences in patterns of marine species richness.

```{r marineRichnessSensTests, results = "markup"}
# Read in the species richness rasters into a raster stack
sppRich <- rast(paste0("AquaMaps/Total_Richness_0", seq(0,5,1), ".tif"))
names(sppRich) <- paste0("Thresh_", seq(0, 0.5, 0.1))

# standardise values to max seen across all threshold
maxR <- max(global(sppRich, max, na.rm = TRUE)[,1])
minR <- min(global(sppRich, min, na.rm = TRUE)[,1])
sppRich_S <- stretch(sppRich, minv = 0, maxv = 1, smin = minR, smax = maxR)

# quantify similarity of species richness predictions using the I statistic
niche_overlap <- Istat(sppRich_S, overlapStat = "I")
niche_overlap[] <- round(niche_overlap, 3)

# estimate of similarity
niche_overlap
```

Estimates of niche similarity show that the native 0.5° resolution from AquaMaps that all estimates of species richness are effectively identical.

## Aggregate the species richness data

Here the species richness data is aggregated to 2.5°x2.5° to match our climate data.

```{r transSppRich}
# aggregate to 2.5° using the mean
sppRich_coarse <- aggregate(x = sppRich[["Thresh_0.4"]], fact = 5, fun = mean, na.rm = TRUE)

# mask to coarse ocean cells only
sppRich_coarse <- mask(sppRich_coarse, ocean_mask)
values(sppRich_coarse) <- round(values(sppRich_coarse))
sppRich_coarse
```

```{r plotSppRich, echo = FALSE, out.width='100%', fig.height = 8, fig.cap="Marine species richness data from Aquamaps"}
xdf <- setDT(as.data.frame(sppRich[["Thresh_0.4"]], xy = TRUE))
xdf[, Res := "Orig"]
xdf_Coarse <- setDT(as.data.frame(sppRich_coarse, xy = TRUE))
xdf_Coarse[, Res := "Coarse"]
xdf <- rbind(xdf, xdf_Coarse)

ggplot() +
  geom_raster(data = xdf, aes(x, y, fill = Thresh_0.4)) +
  facet_wrap(~Res, ncol = 1,
             labeller = as_labeller(x = c(Orig = "Original",
                                          Coarse = "Aggregated"))) +
  scale_fill_gradientn(colours = rocket(100),
                       limits = c(1, 5000),
                       oob = squish,
                       labels = c("Low", "", "High"),
                       breaks = c(1, 2500, 5000),
                       guide = guide_colourbar(
                         title = "Species richness",
                         title.position = "top",
                         title.hjust = 0.5,
                         title.theme = element_text(size = 12, colour = "black"),
                         label.position = "bottom",
                         label.hjust = 0.5,
                         label.theme = element_text(size = 12, colour = "black"),
                         barwidth = 10,
                         barheight = 0.5,
                         nbin = 100,
                         frame.colour = "black",
                         frame.linewidth = 0.5
                       )) +
  geom_sf(data = coastlines, inherit.aes = FALSE) +
  coord_sf(xlim = c(-180,180), ylim = c(-90,90),
           expand = FALSE) +
  theme_void() +
  theme(legend.position = "bottom",
        legend.justification = "center",
        legend.direction = "horizontal",
        strip.text.x = element_text(size = 12, colour = "black"))
```

A quick plot of the values shows the extreme right skewed distribution of both the original and aggregated dataset.

```{r sppRichHist, echo = FALSE, out.width='100%'}
xdf_NAomit <- na.omit(values(sppRich))
xdf_NAomitCoarse <- na.omit(values(sppRich_coarse))
{par(mfrow = c(1, 2), mai = c(0.8,0.5,0.5,0))
dOrig <- density(xdf_NAomit, kernel = "epanechnikov",
             from = 1, to = 6198)
plot(dOrig, 
     ylim = range(dOrig$y),
     xlim = range(xdf_NAomit),
     ylab = "", main = "", axes = FALSE,
     xlab = paste0("\nOriginal", 
                   sprintf("\nmean = %.0f", mean(xdf_NAomit)),
                   sprintf("\nsd = %.2f", sd(xdf_NAomit)),
                   sprintf("\nskewness = %.2f", Skew(xdf_NAomit))))
rug(x = xdf_NAomit, side = 3, col = "#A3A3A368")
axis(2); axis(3); box()
par(mai = c(0.8, 0.1, 0.5, 0.25))
plot(density(xdf_NAomitCoarse, kernel = "epanechnikov",
             from = 1, 
             to = max(xdf_NAomitCoarse)), ylab = "", main = "", axes = FALSE,
     ylim = range(dOrig$y), xlim = range(xdf_NAomit),
     xlab = paste0("\n2.5° aggregated", 
                   sprintf("\nmean = %.0f", mean(xdf_NAomitCoarse)),
                   sprintf("\nsd = %.2f", sd(xdf_NAomitCoarse)),
                   sprintf("\nskewness = %.2f", Skew(xdf_NAomitCoarse))))
rug(x = xdf_NAomitCoarse, side = 3, col = "#A3A3A368")
axis(3); box()}
```

Here we rescale the species richness data to the `{0,1}` range and then export.

```{r rescaleSppRich}
sMax <- global(sppRich_coarse, max, na.rm = TRUE)[,1]
sppRich_rescale <- stretch(sppRich_coarse, minv = 0, maxv = 1, smin = 1, smax = sMax)
writeRaster(sppRich_rescale,
            "outputs/marine_richness_rescaled_wgs84.grd",
            overwrite = FALSE)
```

## Hotspots of marine species richness

Here we extract the locations of cells that are \> 95th percentile of global marine species richness (@ramírez2017), these cells are marine species richness hotspots. The aggregated marine species richness data is used to define the hotspots.

We also produce hotspots at the 90th, and 99th percentile.

```{r locateHotspots, results='hide', message=FALSE}
# Hotspots at 95th percentile
msr_hotspots <- sppRich_coarse
thresh <- quantile(values(msr_hotspots), probs = c(0.90, 0.95, 0.99), na.rm = TRUE)
msr_hotspots[msr_hotspots < thresh["95%"]] <- NA
msr_hotspots[!is.na(msr_hotspots)] <- 1
writeRaster(msr_hotspots,
           "outputs/marine_species_hotspots_wgs84.grd")

# Additional hotspots
msr_90 <- msr_99 <- sppRich_coarse
msr_90[msr_90 < thresh["90%"]] <- NA
msr_99[msr_99 < thresh["99%"]] <- NA
msr_99[!is.na(msr_99)] <- 1; msr_90[!is.na(msr_90)] <- 1

# convert to polygons for plotting and export.
msr_90 <- st_as_sf(as.polygons(msr_90))
msr_95 <- st_as_sf(as.polygons(msr_hotspots))
msr_99 <- st_as_sf(as.polygons(msr_99))

msr_90$Thresh <- 90
msr_95$Thresh <- 95
msr_99$Thresh <- 99

hs_coarse <- rbind(msr_90, msr_95, msr_99)
st_write(hs_coarse[2], "outputs/output_shapes.gpkg", layer = "marine_hotspots",
         quiet = TRUE)
```

The different threshold values used to define hotspots of species richness correspond to values of `r round(thresh[1])` (scaled = `r round(scales::rescale(x = round(thresh[1]), from = c(1, sMax), to = c(0,1)), 2)`), `r round(thresh[2])` (scaled = `r round(scales::rescale(x = round(thresh[2]), from = c(1, sMax), to = c(0,1)), 2)`), and `r round(thresh[3])` (scaled = `r round(scales::rescale(x = round(thresh[3]), from = c(1, sMax), to = c(0,1)), 2)`).

```{r plotCoarseData, echo = FALSE, out.width='100%', fig.cap="Marine species richness data aggregated to 2.5° with marine species richness hotspots overlaid."}
xdf <- as.data.frame(sppRich_rescale, xy = TRUE, na.rm = TRUE)
ggplot() +
  geom_raster(data = xdf, aes(x, y, fill = Thresh_0.4)) +
  scale_fill_gradientn(colours = mako(100, end = 0.9, begin = 0.1, direction = -1),
                       limits = c(0, 1),
                       oob = squish,
                       labels = c("Low", "", "High"),
                       breaks = c(0, 0.5, 1),
                       guide = guide_colourbar(
                         title = "Species richness",
                         title.position = "top",
                         title.hjust = 0.5,
                         title.theme = element_text(size = 12, colour = "black"),
                         label.position = "bottom",
                         label.hjust = 0.5,
                         label.theme = element_text(size = 12, colour = "black"),
                         barwidth = 10,
                         barheight = 0.5,
                         nbin = 100,
                         frame.colour = "black",
                         frame.linewidth = 0.5
                       )) +
  geom_sf(data = coastlines, inherit.aes = FALSE,
          col = "grey30", lwd = 0.5) +
   geom_sf(data = hs_coarse, inherit.aes = FALSE,
          aes(col = as.factor(Thresh)), lwd = 0.25, fill = NA) +
  scale_colour_manual(
    values = c("90" = "#000000", "95" = "#FF0000", "99" = "#FFFF00"),
    guide = guide_legend(
      title = "Hotspot threshold",
      title.position = "top",
      title.hjust = 0.5,
      title.theme = element_text(size = 12, colour = "black"),
      label.position = "bottom",
      label.hjust = 0.5,
      label.theme = element_text(size = 12, colour = "black"),
      keywidth = unit(5, "mm"), 
      keyheight = unit(5, "mm"),
      nrow = 1,
      )) +
  coord_sf(xlim = c(-180,180), ylim = c(-90,90),
           expand = FALSE) +
  theme_void() +
  theme(legend.position = "bottom",
        legend.justification = "center",
        legend.box = "horizontal",
        legend.spacing.x = unit(2, "lines"),
        legend.direction = "horizontal",
        strip.text.x = element_text(size = 12, colour = "black"))
```

## References
