==> Code/03_BivariateMapping.Rmd <==
---
title: "03 - Bivariate Mapping"
author: "Stuart Brown"
output: 
  html_document: 
    theme: flatly
    toc: yes
    toc_depth: 3
    toc_float: true
    fig_width: 7
    fig_height: 5
    fig_caption: yes
    number_sections: no
    keep_md: yes
    df_print: paged
    code_folding: show
    highlight: tango
bibliography: references.bib
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, rows.print = 10, cols.print = 10,
                      rownames.print = FALSE, warning = FALSE,
                      message = FALSE, eval = TRUE, results = "markup", 
                      error = FALSE, strip.white = TRUE,
                      tidy = "styler", dpi = 320, fig.align = "center",
                      strip.white = TRUE, 
                      dev.args = list(png  = list(type = "cairo")))
```

## Read in the data

Here we read in the climate data and species richness data for plotting

```{r importData, results='markup'}
library(data.table)
library(terra)
library(raster)
library(sf)
library(ggplot2)
library(cowplot)
library(rnaturalearth)
library(DescTools)
library(ggpattern)
library(classInt)
library(cptcity)

# Source the bivariate map code
source("source/bivariate_map.R")

source("source/rotate_prj.R")

# Coastlines and aesthetics
coastlines <- ne_coastline(50, returnclass = "sf")
land <- ne_countries(50, returnclass = "sf")
oceans <- read_sf("outputs/output_shapes.gpkg", layer = "ne_ocean")

# Marine realms from Costello et al clipped to oceans
regions <- read_sf("outputs/output_shapes.gpkg", layer = "marine_regions_clipped")

st_agr(oceans) <- "constant"
st_agr(regions) <- "constant"

# 2.5° Ocean mask
ocean_mask <- rast("outputs/ocean_mask.grd")

# Climate data masked to oceans only
rx <- mask(rast("outputs/snr_st_delta_rcp45.grd"), ocean_mask)

# Species richness data
ry <- rast("outputs/marine_richness_rescaled_wgs84.grd")

# hotspot data
rhs <- rast("outputs/marine_species_hotspots_wgs84.grd")
rhs <- st_as_sf(as.polygons(rhs))

# raster stack
rs <- rast(list(rx, ry))
rs <- rs*1 # force into memory
names(rs) <- c("DeltaSNR", "MSR")
rs
```

## Create the data for the bivariate map

The bivariate mapping code has already been sourced from [github](https://gist.github.com/scbrown86/2779137a9378df7b60afd23e0c45c188). Essentially, bivariate maps are a type of thematic map that displays two or more variables on a single map. In this case, we create a bivariate choropleth map that combines the breaks in the values of both datasets into a single multivariate colour scheme.

A quick plot of the values in the rasters suggests that terciles might not be appropriate (see @lucchesi2017 for why terciles are *usually* a good choice), due to the extremely skewed distributions of the datasets.

```{r sppRichHist, echo = FALSE, out.width='100%'}
delCols <- c("#44B360", "#F7900A")
msrCols <- c("#44B360", "#3A88B5")
xdf_Delta <- na.omit(values(rs[["DeltaSNR"]]))
xdf_MSR <- na.omit(values(rs[["MSR"]]))
{par(mfrow = c(1, 2), mai = c(0.8,0.5,0.5,0.1))
dDelta <- density(xdf_Delta, kernel = "epanechnikov",
             from = min(xdf_Delta), to = max(xdf_Delta))
plot(dDelta, 
     ylim = range(dDelta$y),
     xlim = c(-1, 1),
     ylab = "", main = "", axes = FALSE,
     xlab = "Delta SNR")
rug(x = xdf_Delta, side = 3, col = "#A3A3A368")
clBreaks <- classInt::classIntervals(xdf_Delta, n = 3,
                                    style = "quantile")
abline(v = clBreaks$brks[-1],
       col = attr(classInt::findColours(clBreaks, pal = delCols), "palette"))
axis(2); axis(3); box()
dMSR <- density(xdf_MSR, kernel = "epanechnikov",
             from = min(xdf_MSR), to = max(xdf_MSR))
par(mai = c(0.8, 0.5, 0.5, 0.1))
clBreaks <- classInt::classIntervals(xdf_MSR, n = 3,
                                    style = "quantile")
plot(dMSR, ylab = "", main = "", axes = FALSE,
     ylim = range(dMSR$y), xlim = c(0, 1),
     xlab = "Species richness")
rug(x = xdf_MSR, side = 3, col = "#A3A3A368")
abline(v = clBreaks$brks[-1],
       col = attr(classInt::findColours(clBreaks, pal = msrCols), "palette"))
axis(2); axis(3); box()}
```

A more suitable classification may come from using Fishers clustering algorithm (@fisher1958, @wand1997). The black dotted line corresponds to the scaled 95th percentile used to identify hotspots of species richness.

```{r sppRichHistJenks, echo = FALSE, fig.height = 5, out.width='100%'}
nClass <- 4
delCols <- c("#44B360", "#F7900A")
msrCols <- c("#44B360", "#3A88B5")
{par(mfrow = c(1, 2), mai = c(0.8,0.5,0.5,0.1))
plot(dDelta, 
     ylim = range(dDelta$y),
     xlim = c(-1, 1),
     ylab = "", main = "", axes = FALSE,
     xlab = "Delta SNR")
rug(x = xdf_Delta, side = 3, col = "#A3A3A368")
clBreaks <- {set.seed(54);classInt::classIntervals(xdf_Delta, n = nClass,
                                         style = "fisher", samp_prop = 1)}
abline(v = clBreaks$brks[-1],
       col = attr(classInt::findColours(clBreaks, pal = delCols), "palette"))
axis(2); axis(3); box()
par(mai = c(0.8, 0.5, 0.5, 0.1), mgp = c(1, 1, 0))
plot(dMSR, ylab = "", main = "", axes = FALSE,
     xlab = "Species richness",
     ylim = range(dMSR$y), xlim = c(0, 1))
rug(x = xdf_MSR, side = 3, col = "#A3A3A368")
clBreaks <- {set.seed(54);classInt::classIntervals(xdf_MSR, n = nClass,
                                         style = "fisher", samp_prop = 1)}
abline(v = clBreaks$brks[-1],
       col = attr(classInt::findColours(clBreaks, pal = msrCols), "palette"))
abline(v = 0.19, col = "black", lty = 2)
axis(2); axis(3); box()
par(mfrow = c(1, 1))}
```

Make the bivariate map between Delta SNR and species richness

```{r makeBiVark, results='hide', fig.show='hide', eval = TRUE}
# Create the colour matrix
col.matrix <- colmat(nbreaks = 4, 
                     breakstyle = "fisher",
                     xlab = "delta SNR", 
                     ylab = "spp. rich",
                     bottomleft = "#c3b3d8", upperleft = "#240d5e",
                     bottomright = "#ffcc80", upperright = "#b30000",
                     saveLeg = FALSE, plotLeg = TRUE,
                     ## sampling argument to classInt::classIntervals
                     samp_prop = 1)

# create the bivariate raster
bivmap <- bivariate.map(rasterx = as(rs[["DeltaSNR"]], "Raster"),
                        rastery = as(rs[["MSR"]], "Raster"),
                        export.colour.matrix = FALSE,
                        colourmatrix = col.matrix,
                        ## sampling argument to classInt::classIntervals
                        samp_prop = 1)

# Convert to data.table for plotting
bivMapDF <- setDT(as.data.frame(bivmap, xy = TRUE))
colnames(bivMapDF)[3] <- "BivValue"
bivMapDF <- melt(bivMapDF, id.vars = c("x", "y"),
                 measure.vars = "BivValue",
                 value.name = "bivVal",
                 variable.name = "Variable")
```

A very quick plot of the relationship between Delta SNR and marine species richness, shows the following.

```{r bivPlot, echo = FALSE, out.width='100%', fig.cap="Delta SNR vs Marine Species richness", eval = TRUE}
map <- ggplot() +
  geom_raster(data = bivMapDF, aes(x = x, y = y, fill = bivVal),
              show.legend = FALSE) +
  geom_sf(data = rhs, inherit.aes = FALSE, colour = "white", fill = NA) +
  geom_sf(data = land, inherit.aes = FALSE, colour = NA, fill = "grey85") +
  geom_sf(data = regions, inherit.aes = FALSE, fill = NA, colour = "grey40") +
  geom_sf(data = coastlines, inherit.aes = FALSE, fill = "#E9E9E9", colour = "black") +
  scale_fill_gradientn(colours = col.matrix, na.value = "transparent") +
  theme_void() +
  coord_sf(expand = FALSE)
fig <- ggdraw(map) +
  draw_plot(BivLegend +
              theme(plot.background = element_rect(fill = "white", colour = NA)),
            width = 0.25, height = 0.25, x = 0.40, y = 0.05, halign = 0.5)
fig
```

We can now make some latitudinal histograms of Delta SNR and species richness

```{r latHisto, echo = TRUE, eval = TRUE, results="markdown"}
zones <- init(rs, fun = "y")
zonal_summary <- setDT(zonal(rs, zones, mean, na.rm = TRUE))
setnafill(zonal_summary, type = "const", fill = 0)
setorder(zonal_summary, -lyr1, na.last = TRUE)
names(zonal_summary) <-  c("Zone", "Delta", "SppRich")
summary(zonal_summary)

hist_right <- ggplot(data = zonal_summary) +
  geom_col(aes(x = Zone, y = Delta, fill = Delta),
           width = rep(as.numeric(abs((zonal_summary[1,1] - zonal_summary[2,1]))),
                       nrow(zonal_summary))) +
  scale_y_continuous(
    breaks = c(-0.20, 0.0, 0.20),
    limits = c(-0.30, 0.30),
    labels = c("Neg.", expression(atop("\n",paste(delta, " SNR"))), "Pos."),
    oob = scales::squish
    ) +
  scale_fill_gradientn(colours = cptcity::cpt("ncl_BlueYellowRed", 100),
                       na.value = "transparent") +
  coord_flip(expand = FALSE) +
  theme_minimal() +
  geom_hline(yintercept = 0) +
  labs(title = NULL, x = NULL, y = NULL) +
  theme(legend.position = "none", axis.text.y = element_blank(),
        axis.text.x = element_text(colour = "black", size = 10),
        axis.ticks = element_blank(), panel.grid = element_blank())
hist_right

hist_left <- ggplot(zonal_summary) +
  # negate values so they are on the left side of the axis
  geom_col(aes(x = Zone, y = -SppRich, fill = -SppRich),
           width = rep(as.numeric(abs((zonal_summary[1,1] - zonal_summary[2,1]))),
                       nrow(zonal_summary))) +
  scale_y_continuous(
    breaks = c(-0.16,-0.085, -0.01),
    labels = c("High", expression(atop("\n","Species richness")), "Low")) +
  scale_fill_gradientn(colours = viridis::mako(100), 
                       na.value = "transparent") +
  coord_flip(expand = FALSE) +
  theme_minimal() +
  geom_hline(yintercept = 0) +
  labs(title = NULL, x= NULL, y = NULL) +
  theme(legend.position = "none", axis.text.y = element_blank(),
        axis.text.x = element_text(colour = "black", size = 10),
        axis.ticks = element_blank(), panel.grid = element_blank())
hist_left
```

Here we'll remake the density plots that show the breaks that are used to define the two axes in the bivariate map.

```{r bivMapHist, results='markup', echo = FALSE}
dDeltaDF <- data.frame(x = dDelta$x, y = dDelta$y)
# SNR breaks
classIntervals(dDeltaDF$x, n = 4, style = "fisher", samp_prop = 1)
snr_dens <- ggplot(dDeltaDF, aes(x = x)) +
  geom_line(aes(y = y, colour = x), show.legend = FALSE, lwd = 1,
            lineend = "square", linejoin = "bevel") +
  scale_x_continuous(limits = c(-1,1), expand = c(0.01,0.01)) +
  scale_colour_gradientn(
    colours = findColours(classIntervals(dDeltaDF$x, n = 4,
                                         style = "fisher",
                                         samp_prop = 1), 
                          pal = c("#c3b3d8","#ffcc80"))) +
  scale_y_continuous() +
  geom_rug(data = as.data.frame(xdf_Delta),
           aes(x = DeltaSNR),
           inherit.aes = FALSE, col = "#A6A6A652", sides = "b") +
  theme_half_open() +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank()) +
  labs(x = expression(paste(delta, " SNR")),
       y = "density")
snr_dens
dMSRDF <- data.frame(x = dMSR$x, y = dMSR$y)
classIntervals(dMSRDF$x, n = 4, style = "fisher", samp_prop = 1)
msr_dens <- ggplot(dMSRDF, aes(x = x)) +
  geom_line(aes(y = y, colour = x), show.legend = FALSE, lwd = 1,
            lineend = "square", linejoin = "bevel") +
  scale_x_continuous(limits = c(0,1), expand = c(0.01,0.01)) +
  scale_colour_gradientn(
    colours = findColours(classIntervals(dMSRDF$x, n = 4,
                                         style = "fisher",
                                         samp_prop = 1), 
                          pal = c("#c3b3d8", "#240d5e"))) +
  scale_y_continuous() +
  geom_vline(xintercept = 0.19, lty = 2, lwd = 0.1) +
  geom_rug(data = as.data.frame(xdf_MSR),
           aes(x = MSR),
           inherit.aes = FALSE, col = "#A6A6A652", sides = "b") +
  theme_half_open() +
  theme(axis.text.y = element_blank(), axis.ticks.y = element_blank()) +
  labs(x = "MSR",
       y = "density")
msr_dens
```

## Make Figure 1

Here we combine the various plots to produce figure 1. As part of making figure 1, the bivariate map is converted to a Mollweide equal area projection, with a central longitude of 160°

```{r makeFig1, fig.height=4, fig.width=10, out.width='100%', fig.cap="Figure 01 - Bivariate map of delta SNR vs marine species richness (a). Histograms show the latitundinal variation in marine species richness (b) and delta SNR (c). Inset plots show density plots of marine species richess (d) and delta SNR (e). Colours on plots d and e correpsond to the colours along the axes of the bivariate plot, with changes in colours showing the location of break points.", results='hide', fig.show='hold', echo = FALSE}
mollPrj <- 'PROJCS["Pacific_Centrered_Mollweide",
 GEOGCS["GCS_WGS_1984",
  DATUM["D_WGS_1984",
   SPHEROID["WGS_1984",6378137.0,298.257223563]],
  PRIMEM["Greenwich",0.0],
  UNIT["Degree",0.0174532925199433]],
 PROJECTION["Mollweide"],
 PARAMETER["False_Easting",0.0],
 PARAMETER["False_Northing",0.0],
 PARAMETER["Central_Meridian",160],
 UNIT["Meter",1.0]]'

templ_ras <- rast(resolution = 2.5, crs = "EPSG:4326")
templ_ras <- project(templ_ras, y = mollPrj)
res(templ_ras) <- c(100000) # 100km only for plotting

# Projected bivariate map
bivProj <- project(as(bivmap, "SpatRaster"), y = templ_ras, method = "near")
bivProj <- setDT(as.data.frame(bivProj, xy = TRUE))
colnames(bivProj)[3] <- "BivValue"
bivProj <- melt(bivProj, id.vars = c("x", "y"),
                 measure.vars = "BivValue",
                 value.name = "bivVal",
                 variable.name = "Variable")

# projected shapefiles for aesthetics
hotspots_prj <- read_sf("outputs/output_shapes.gpkg", layer = "marine_hotspots")
hotspots_prj <- hotspots_prj[hotspots_prj$Thresh == 95, ]
hotspots_prj <- rotate_prj(hotspots_prj, crs = st_crs(templ_ras))

land_prj <- read_sf("shapes/mapping_shapes.gpkg", layer = "land")
regions_prj <- read_sf("shapes/mapping_shapes.gpkg", layer = "regions_merged")
coast_prj <- read_sf("shapes/mapping_shapes.gpkg", layer = "coast")
bbox_prj <- read_sf("shapes/mapping_shapes.gpkg", layer = "bbox")

# fig 1 map
map01 <- ggplot() +
  geom_raster(data = bivProj, aes(x = x, y = y, fill = bivVal),
              show.legend = FALSE) +
  geom_sf(data = land_prj, size = 0.25,
          inherit.aes = FALSE, colour = NA, fill = "#E9E9E9") +
  geom_sf(data = regions_prj, size = 0.25,
          inherit.aes = FALSE, fill = NA, colour = "grey40") +
   geom_sf_pattern(data = hotspots_prj, size = 0.25,
                   pattern = "crosshatch", pattern_size = 0.25,
                   pattern_density = 0.05, pattern_spacing = 0.01,
                   pattern_colour  = "white",
                   inherit.aes = FALSE, colour = "white", fill = NA) +
  geom_sf(data = coast_prj, size = 0.25,
          inherit.aes = FALSE, fill = "#E9E9E9", colour = "black") +
  geom_sf(data = bbox_prj, size = 0.25,
          inherit.aes = FALSE, fill = NA, colour = "black") +
  scale_fill_gradientn(colours = col.matrix, na.value = "transparent") +
  theme_map() +
  coord_sf(expand = FALSE)

fig01 <- plot_grid(hist_left, map01, hist_right, align = "h", axis = "tb", nrow = 1, 
          rel_widths = c(0.20, 1, 0.20), rel_heights = c(1,1,1),
          labels = c("b", "a", "c"), label_size = 14) +
  draw_plot(BivLegend +
              scale_x_continuous(expand = c(0.02, 0.02)) +
              scale_y_continuous(expand = c(0.02, 0.02)) +
              coord_fixed(ratio = 1, clip = "off") +
              theme(plot.background = element_rect(fill = NA, colour = NA),
                    axis.title.x = element_text(size = 8, vjust = -0.5, hjust = 0.5),
                    axis.title.y = element_text(size = 8, vjust = 1)) +
              labs(x = expression(paste(delta, " SNR")), 
                    y = "MSR") +
              annotate("segment", x = 1, xend = 4, y = 0.15, yend = 0.15,
                       arrow = arrow(ends = "last", angle = 30, 
                                     length = unit(.2,"cm"))) +
              annotate("segment", x = 0.10, xend = 0.10, y = 1, yend = 4,
                       arrow = arrow(ends = "last", angle = 30, 
                                     length = unit(.2,"cm"))),
            width = 0.2125, height = 0.2125, x = 0.280, y = 0.80, halign = 0.5,
            vjust = 0.5)
fig01 <- fig01 +
  draw_plot(snr_dens +
    scale_y_continuous(expand = c(0.15,0.15),
                       sec.axis = sec_axis(~., name = "density")) +
    theme(plot.background = element_rect(fill = NA, colour = NA),
          axis.text = element_text(size = 8),
          axis.title = element_text(size = 8),
          axis.line = element_line(size = 0.25, colour = "black"),
          axis.ticks.x = element_line(size = 0.25, colour = "black"),
          axis.title.y.left = element_blank(),
          axis.title.y.right = element_text(angle = 90, vjust = 2),
          axis.line.y.left = element_blank()), 
            width = 0.25, height = 0.32, x = 0.635, y = -0.0285,
            halign = 0.5, valign = 0.5) +
  draw_plot_label(label = "e", x = 0.860, y = 0.30, size = 14)
fig01 <- fig01 +
  draw_plot(msr_dens +
              scale_y_continuous(expand = c(0.15,0.15)) +
               theme(plot.background = element_rect(fill = NA, colour = NA),
                    axis.text = element_text(size = 8),
                    axis.title = element_text(size = 8),
                    axis.line = element_line(size = 0.25, colour = "black"),
                    axis.ticks.x = element_line(size = 0.25, colour = "black")), 
            width = 0.25, height = 0.32, x = 0.15, y = -0.0285,
            halign = 0.5, valign = 0.5) +
  draw_plot_label(label = "d", x = 0.15, y = 0.30, size = 14)

# save to PDF
ggsave(filename = "Figure01.pdf", plot = fig01,
       device = cairo_pdf, path = "Figures",
       width = 10, height = 4, units = "in", dpi = 350)

# save to PNG for MS
ggsave(filename = "Figure01.png", plot = fig01,
       device = "png", type = "cairo", path = "Figures",
       bg = "white", width = 10, height = 4, units = "in", dpi = 350)

# plot
fig01
```

## References
